add funtion for Authorizations with POST method.
